pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "irsam2-benchmark"
version = "0.1.0"
description = "Paper-grade SAM2 benchmark platform for infrared transfer, adaptation, distillation, quantization, and evaluation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.5.1",
  "torchvision>=0.20.1",
  "numpy>=1.24.4",
  "Pillow>=9.4.0",
  "tqdm>=4.66.1",
  "hydra-core>=1.3.2",
  "iopath>=0.1.10",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]

[project.scripts]
irsam2-benchmark = "irsam2_benchmark.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
